refactor(common): move AwsClients to config package and update imports accordingly
